main_django_microservice/pyproject.toml
[tool.poetry]
name = "main_django_microservice"
version = "0.1.0"
description = ""
authors = ["ku-baha <kulbarakovbh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
pika = "^1.3.2"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"


[tool.poetry.group.deploy.dependencies]
gunicorn = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
